fixed user profile creating upon sign up, changed some styling to django accounts templates
