+++ cafeOrdering/cafeOrdering/urls.py
@@ -16,11 +16,12 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from users.views import sign_up
+from users.views import sign_up, view_profile
 
 urlpatterns = [
     path('', include('homepage.urls')),
     path('accounts/signup/', sign_up, name='account_signup'),
     path('accounts/', include('allauth.urls')),
     path('admin/', admin.site.urls),
+    path('profile/', view_profile, name='view_profile'),
 ]
