+++ cafeOrdering/templates/users/profile.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %}
+
+{% block content %}
+<div>
+    <h3>Profile</h3>
+</div>
+{% endblock %}
