+++ cafeOrdering/users/apps.py
@@ -1,6 +1,8 @@
 from django.apps import AppConfig
 
-
 class UsersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'users'
+
+    def ready(self):
+        import users.signals
