+++ cafeOrdering/users/urls.py
@@ -2,5 +2,5 @@ from django.urls import path
 from .views import view_profile
 
 urlpatterns = [
-    path('profile/', view_profile, name='view_profile'),
+    
 ]
